Fix issue with implements clause and imported interfaces (#3)
## Summary:
The logic that flowgen was using to determine whether a heritage clause was an 'extends' or a 'implements' clause was faulty.  I've refactor it to look at clause.token which contains either an 'extends' or 'implements' token.  This simplify the code a bunch and fixed the issue that the tool was running into in wonder-blocks where it thought something part of the 'extends' clause when actually it was part of the 'implements' clause.

Issue: None

## Test plan:
- yarn test

Author: kevinbarabash

Reviewers: jeresig

Required Reviewers:

Approved By: jeresig

Checks:

Pull Request URL: https://github.com/Khan/flowgen/pull/3